# docs/swagger.yaml
basePath: /api
definitions:
...
  entity.Customer:
    properties:
      alamat:
        items:
          $ref: '#/definitions/entity.CustomerAddress'
        type: array
      customer_class:
        type: string
      customer_code:
        type: string
      customer_create_at:
        type: string
      customer_dob:
        $ref: '#/definitions/sql.NullString'
      customer_email:
        type: string
      customer_gender:
        type: string
      customer_id:
        type: string
      customer_last_status:
        type: integer
      customer_major_id:
        type: string
      customer_name:
        type: string
      customer_password:
        type: string
      customer_phonenumber:
        type: string
      customer_profile_pic:
        type: string
      customer_profile_pic_path:
        type: string
      customer_status:
        type: integer
      customer_update_at:
        type: string
      jurusan:
        $ref: '#/definitions/entity.Major'
    type: object
  entity.CustomerAddress:
    properties:
      address_city:
        type: string
      address_city_id:
        type: string
      address_create_at:
        type: string
      address_customer_id:
        type: string
      address_district:
        type: string
      address_district_id:
        type: string
      address_id:
        type: string
      address_main:
        type: integer
      address_name:
        type: string
      address_postal_code:
        type: string
      address_province:
        type: string
      address_province_id:
        type: string
      address_text:
        type: string
      address_update_at:
        type: string
      address_village:
        type: string
      address_village_id:
        type: string
    type: object
...
  entity.CustomerResponse:
    properties:
      alamat:
        items:
          $ref: '#/definitions/entity.CustomerAddress'
        type: array
      customer_class:
        type: string
      customer_code:
        type: string
      customer_create_at:
        type: string
      customer_dob:
        $ref: '#/definitions/sql.NullString'
      customer_email:
        type: string
      customer_gender:
        type: string
      customer_id:
        type: string
      customer_last_status:
        type: integer
      customer_major_id:
        type: string
      customer_name:
        type: string
      customer_phonenumber:
        type: string
      customer_profile_pic:
        type: string
      customer_profile_pic_path:
        type: string
      customer_status:
        type: integer
      customer_update_at:
        type: string
      jurusan:
        $ref: '#/definitions/entity.MajorResponse'
    type: object
...
  entity.Login:
    properties:
      user_device_metadata:
        type: string
      user_email:
        type: string
      user_fcmtoken:
        type: string
      user_id:
        type: string
      user_password:
        type: string
    required:
    - user_email
    - user_password
    type: object
  entity.Major:
    properties:
      major_id:
        type: string
      major_name:
        type: string
    type: object
  entity.MajorResponse:
    properties:
      major_id:
        type: string
      major_name:
        type: string
    type: object
...
  entity.Pegawai:
    properties:
      pegawai_code:
        type: string
      pegawai_gender:
        type: string
      pegawai_id:
        type: string
      pegawai_name:
        type: string
      pegawai_phonenumber:
        type: string
    type: object
...
  entity.Transaction:
    properties:
      customer:
        $ref: '#/definitions/entity.CustomerResponse'
      trans_create_at:
        type: string
      trans_customer_id:
        type: string
      trans_detail:
        items:
          $ref: '#/definitions/entity.TransactionDetail'
        type: array
      trans_discount:
        type: number
      trans_id:
        type: string
      trans_invoice:
        type: string
      trans_order_id:
        type: string
      trans_product_total:
        type: integer
      trans_qty_total:
        type: integer
      trans_received_total:
        type: number
      trans_refund_total:
        type: number
      trans_status:
        type: integer
      trans_subtotal:
        type: number
      trans_total:
        type: number
      trans_user_id:
        type: string
      user:
        $ref: '#/definitions/entity.User'
    type: object
  entity.TransactionDetail:
    properties:
      trans_detail_id:
        type: string
      trans_detail_parent_id:
        type: string
      trans_detail_price:
        type: number
      trans_detail_product_varian_id:
        type: string
      trans_detail_qty:
        type: integer
      trans_detail_subtotal:
        type: number
    type: object
  entity.User:
    properties:
      pegawai:
        $ref: '#/definitions/entity.Pegawai'
      user_email:
        type: string
      user_has_mobile_access:
        type: integer
      user_id:
        type: string
      user_name:
        type: string
      user_password:
        type: string
      user_pegawai_id:
        type: string
      user_role_id:
        type: string
    type: object
...
  handler.WebResponse:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
    type: object
  sql.NullString:
    properties:
      string:
        type: string
      valid:
        description: Valid is true if String is not NULL
        type: boolean
    type: object
...
host: 127.0.0.1:3000
info:
  contact:
    email: support@ecanteen.com
    name: API Support
  description: API documentation for E-Canteen Cashier System - A comprehensive canteen
    management solution
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: E-Canteen Cashier API
  version: 1.2.0
paths:
  /customer:
    get:
      description: Retrieve all customers
      produces:
      - application/json
      responses:
        "200":
          description: Customers list
          schema:
            allOf:
            - $ref: '#/definitions/handler.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.Customer'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get all customers
      tags:
      - Customer
    post:
      consumes:
      - application/json
      description: Register a new customer
      parameters:
      - description: Customer data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.Customer'
      produces:
      - application/json
      responses:
        "201":
          description: Customer registered
          schema:
            allOf:
            - $ref: '#/definitions/handler.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.Customer'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/handler.WebResponse'
      summary: Register customer
      tags:
      - Customer
  /customer/{customerId}:
    delete:
      description: Delete customer by ID
      parameters:
      - description: Customer ID
        in: path
        name: customerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Customer deleted
          schema:
            $ref: '#/definitions/handler.WebResponse'
        "404":
          description: Customer not found
          schema:
            $ref: '#/definitions/handler.WebResponse'
      security:
      - BearerAuth: []
      summary: Delete customer
      tags:
      - Customer
    get:
      description: Retrieve customer details by ID
      parameters:
      - description: Customer ID
        in: path
        name: customerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Customer details
          schema:
            allOf:
            - $ref: '#/definitions/handler.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.Customer'
              type: object
        "404":
          description: Customer not found
          schema:
            $ref: '#/definitions/handler.WebResponse'
      security:
      - BearerAuth: []
      summary: Get customer by ID
      tags:
      - Customer
    put:
      consumes:
      - application/json
      description: Update customer information
      parameters:
      - description: Customer ID
        in: path
        name: customerId
        required: true
        type: string
      - description: Customer data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.Customer'
      produces:
      - application/json
      responses:
        "200":
          description: Customer updated
          schema:
            allOf:
            - $ref: '#/definitions/handler.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.Customer'
              type: object
        "404":
          description: Customer not found
          schema:
            $ref: '#/definitions/handler.WebResponse'
      security:
      - BearerAuth: []
      summary: Update customer
      tags:
      - Customer
  /customer/login:
    post:
      consumes:
      - application/json
      description: Authenticate customer and get JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.Login'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with JWT token
          schema:
            allOf:
            - $ref: '#/definitions/handler.WebResponse'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/handler.WebResponse'
      summary: Customer login
      tags:
      - Customer
  /customer/logout:
    post:
      consumes:
      - application/json
      description: Logout customer and invalidate token
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/handler.WebResponse'
      security:
      - BearerAuth: []
      summary: Customer logout
      tags:
      - Customer
...
  /kasir/login:
    post:
      consumes:
      - application/json
      description: Authenticate kasir user and get JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.Login'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with JWT token
          schema:
            allOf:
            - $ref: '#/definitions/handler.WebResponse'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/handler.WebResponse'
      summary: User login (Kasir)
      tags:
      - Authentication
  /kasir/logout:
    put:
      consumes:
      - application/json
      description: Logout kasir user and invalidate token
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/handler.WebResponse'
      security:
      - BearerAuth: []
      summary: User logout (Kasir)
      tags:
      - Authentication
  /kasir/transaction:
    post:
      consumes:
      - application/json
      description: Create a new cashier transaction
      parameters:
      - description: Transaction data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.Transaction'
      produces:
      - application/json
      responses:
        "201":
          description: Transaction created
          schema:
            allOf:
            - $ref: '#/definitions/handler.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.Transaction'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/handler.WebResponse'
      security:
      - BearerAuth: []
      summary: Create transaction
      tags:
      - Transactions
  /kasir/transaction/{transId}:
    get:
      description: Retrieve transaction details by ID
      parameters:
      - description: Transaction ID
        in: path
        name: transId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction details
          schema:
            allOf:
            - $ref: '#/definitions/handler.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.Transaction'
              type: object
        "404":
          description: Transaction not found
          schema:
            $ref: '#/definitions/handler.WebResponse'
      security:
      - BearerAuth: []
      summary: Get transaction by ID
      tags:
      - Transactions
  /kasir/transaction_detail:
    get:
      description: Retrieve transaction detail items
      parameters:
      - description: Transaction ID
        in: query
        name: trans_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction details
          schema:
            allOf:
            - $ref: '#/definitions/handler.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.TransactionDetail'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get transaction details
      tags:
      - Transactions
...
  /transaction:
    get:
      description: Retrieve all transactions
      produces:
      - application/json
      responses:
        "200":
          description: Transactions list
          schema:
            allOf:
            - $ref: '#/definitions/handler.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.Transaction'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get all transactions
      tags:
      - Transactions
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"